Toilet scale en positie verbeterd

+++ src/Application/Simulation/Simulation.java
@@ -68,7 +68,7 @@ public class Simulation extends Application {
     public void init() throws Exception {
         camera = new Camera();
         map = new Map("/FestivalMap.json", this.pathFinder);
-        this.toilet = new Toilet(new Point2D.Double(0,0));
+        this.toilet = new Toilet(new Point2D.Double(282,4312));
 //        for (int i = 0; i < 3; i++) {
 //            Visitor visitor = new Visitor(new Point2D.Double(Math.random()*(128*8), Math.random()*(128*8)),1);
 //            visitors.add(visitor);
